Agent: Rename credential_generation -> credential_generators
