+++ monkey/infection_monkey/exploit/powershell.py
@@ -14,7 +14,7 @@ from infection_monkey.exploit.consts import WIN_ARCH_32, WIN_ARCH_64
 from infection_monkey.exploit.HostExploiter import HostExploiter
 from infection_monkey.exploit.powershell_utils import utils
 from infection_monkey.exploit.powershell_utils.auth_options import AuthOptions
-from infection_monkey.exploit.powershell_utils.credential_generation import get_credentials
+from infection_monkey.exploit.powershell_utils.credential_generators import get_credentials
 from infection_monkey.exploit.powershell_utils.utils import (
     IClient,
     get_client_based_on_auth_options,
