+++ monkey/infection_monkey/exploit/powershell_utils/credential_generators.py
