+++ monkey/tests/unit_tests/infection_monkey/exploit/powershell_utils/test_utils.py
@@ -1,6 +1,6 @@
 from infection_monkey.exploit.powershell_utils import utils
 from infection_monkey.exploit.powershell_utils.auth_options import AuthOptions
-from infection_monkey.exploit.powershell_utils.credential_generation import get_credentials
+from infection_monkey.exploit.powershell_utils.credential_generators import get_credentials
 from infection_monkey.model.host import VictimHost
 
 TEST_USERNAMES = ["user1", "user2"]
